refactor modal
