+++ src/components/widget-edit.js
@@ -1,5 +1,4 @@
 import React, {createElement as el} from 'react'
-import ButtonOutline from 'rebass/dist/ButtonOutline'
 
 import AttributionEditor from './attribution-editor'
 import WidgetUnsupported from './widget-unsupported'
@@ -15,41 +14,9 @@ const COMPONENTS = {
   unsupported: WidgetUnsupported,
 }
 
-function stopPropagation (event) { event.stopPropagation() }
-
 
 class WidgetEdit extends React.Component {
   render () {
-    return el('div',
-      {
-        style: {
-          padding: '1rem',
-          backgroundColor: 'white',
-          maxWidth: 720,
-          margin: '1rem auto 4rem',
-          border: '1px solid silver',
-          borderRadius: 2,
-        },
-        onClick: stopPropagation,
-      },
-      el('div',
-        {
-          style: {
-            textAlign: 'right',
-            marginBottom: '1rem',
-          },
-        },
-        el(ButtonOutline,
-          {
-            onClick: this.props.onClose,
-          },
-          'Close'
-        ),
-      ),
-      this.renderEditor(),
-    )
-  }
-  renderEditor () {
     const {initialBlock, coverPrefs} = this.props
     const {id, type} = initialBlock
     const Component = COMPONENTS[type] || COMPONENTS.unsupported
@@ -59,7 +26,7 @@ class WidgetEdit extends React.Component {
 WidgetEdit.propTypes = {
   initialBlock: React.PropTypes.object.isRequired,
   coverPrefs: React.PropTypes.object.isRequired,
-  onClose: React.PropTypes.func.isRequired,
+  // onClose: React.PropTypes.func.isRequired,
 }
 WidgetEdit.contextTypes = {
   store: React.PropTypes.object,
